+++ libfreerdp-utils/args.c
@@ -127,10 +127,10 @@ int freerdp_parse_args(rdpSettings* settings, int argc, char** argv,
 				printf("missing width\n");
 				return 0;
 			}
-			settings->width = strtol(argv[index], &p, 10);
+			settings->width = (uint16) strtol(argv[index], &p, 10);
 			if (*p == 'x')
 			{
-				settings->height = strtol(p + 1, &p, 10);
+				settings->height = (uint16) strtol(p + 1, &p, 10);
 			}
 			else
 			{
